+++ lib/datacite/mapping/module_info.rb
@@ -4,7 +4,7 @@ module Datacite
     NAME = 'datacite-mapping'
 
     # The version of this gem
-    VERSION = '0.1.18'
+    VERSION = '0.2.0'
 
     # The copyright notice for this gem
     COPYRIGHT = 'Copyright (c) 2016 The Regents of the University of California'